Conflicts:
	Core/WCell.Constants/AreaTriggers/AreaTriggerId.cs
	Core/WCell.Constants/Factions/FactionId.cs
	Core/WCell.Constants/Factions/FactionReputationIndex.cs
	Core/WCell.Constants/Factions/FactionTemplateId.cs
	Core/WCell.Constants/GameObjects/GOEntryId.cs
	Core/WCell.Constants/Items/ItemId.cs
	Core/WCell.Constants/Items/ItemSetId.cs
	Core/WCell.Constants/Items/TotemCategory.cs
	Core/WCell.Constants/NPCs/CreatureFamilyId.cs
	Core/WCell.Constants/NPCs/NPCId.cs
	Core/WCell.Constants/Realm/RealmCategory.cs
	Core/WCell.Constants/Skills/SkillId.cs
	Core/WCell.Constants/Spells/Range.cs
	Core/WCell.Constants/Spells/SpellFailedReason.cs
	Core/WCell.Constants/Spells/SpellFocus.cs
	Core/WCell.Constants/Spells/SpellId.cs
	Core/WCell.Constants/Spells/SpellLineId.cs
	Core/WCell.Constants/Spells/SpellMechanic.cs
	Core/WCell.Constants/Talents/TalentId.cs
	Core/WCell.Constants/Talents/TalentTreeId.cs
	Core/WCell.Constants/Updates/UpdateFieldEnums.cs
	Core/WCell.Constants/Updates/UpdateFields.cs
	Core/WCell.Constants/World/MapId.cs
	Core/WCell.Constants/World/ZoneId.cs
	Services/WCell.RealmServer/Global/World.DBC.cs
	Services/WCell.RealmServer/Global/ZoneInfo.cs
	Services/WCell.RealmServer/NPCs/SpawnPoint.cs
	Services/WCell.RealmServer/Spells/Spell.DBC.cs
	Services/WCell.RealmServer/Spells/SpellLines.Def.cs
	Services/WCell.RealmServer/Talents/TalentMgr.cs
	Services/WCell.RealmServer/Talents/Talents.DBC.cs
	Utilities/WCell.MPQTool/DBCTool.cs

+++ Core/WCell.Constants/Updates/UpdateFieldMgr.cs
@@ -10,7 +10,7 @@ namespace WCell.Constants.Updates
 		public static readonly UpdateFieldCollection[] Collections = new UpdateFieldCollection[UpdateField.ObjectTypeCount];
 		public static readonly ObjectTypeId[] InheritedTypeIds = new ObjectTypeId[UpdateField.ObjectTypeCount];
 
-		public static uint ExplorationZoneFieldSize;
+		public static int ExplorationZoneFieldSize;
 
 		public static void Init()
 		{
@@ -80,7 +80,7 @@ namespace WCell.Constants.Updates
 		private static void FixFields()
 		{
 			//ExplorationZoneFieldSize = Get(ObjectTypeId.Player).Fields[(uint)PlayerFields.EXPLORED_ZONES_1].Size;
-			ExplorationZoneFieldSize = UpdateFields.AllFields[(int)ObjectTypeId.Player][(uint)PlayerFields.EXPLORED_ZONES_1].Size;
+			ExplorationZoneFieldSize = (int) UpdateFields.AllFields[(int)ObjectTypeId.Player][(uint)PlayerFields.EXPLORED_ZONES_1].Size;
 		}
 
 		private static void InitInheritance()
